@@ -47,7 +47,7 @@ class ImageApiMixin:
             image (str): The image to show history for
 
         Returns:
-            (str): The history of the image
+            (list): The history of the image
 
         Raises:
             :py:class:`docker.errors.APIError`
